Refactor: Break utils into separate files
